Refactored code file structure.  Partly refactored docs.
