+++ scintillometry/__init__.py
@@ -27,3 +27,5 @@ if (sys.version_info < tuple(
 if not _ASTROPY_SETUP_:
     # For egg_info test builds to pass, put package imports here.
     pass
+
+from .channelize import ChannelizeTask
