+++ scintillometry/base.py
