--- scintillometry/channelize/__init__.py
@@ -1,4 +1,0 @@
-# Licensed under the GPLv3 - see LICENSE
-"""Channelization modules."""
-
-from .base import ChannelizeTask
