--- scintillometry/dispersion/__init__.py
@@ -1,3 +1,0 @@
-# Licensed under the GPLv3 - see LICENSE
-"""Scintilllometry dispersion module."""
-from .dm import DispersionMeasure
